update auditoria logs

+++ backend/src/main/java/com/example/backend/auditoria/logAcoes/LogAcoesRepository.java
@@ -2,5 +2,5 @@ package com.example.backend.auditoria.logAcoes;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LogAcoesRepository extends JpaRepository<LogAcoes, Integer> {
+public interface LogAcoesRepository extends JpaRepository<LogAcoes, Long> {
 }
