+++ backend/src/main/java/com/example/backend/auditoria/logErros/LogErrosRepository.java
@@ -2,5 +2,5 @@ package com.example.backend.auditoria.logErros;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LogErrosRepository extends JpaRepository<LogErros, Integer> {
+public interface LogErrosRepository extends JpaRepository<LogErros, Long> {
 }
